fix: promotion to master

+++ core/src/main/java/it/gov/pagopa/rtd/transaction_filter/service/StoreService.java
@@ -75,7 +75,7 @@ public interface StoreService {
      * @param posType the transaction POS type
      * @return boolean true if dirty data has been found during aggregation
      */
-    boolean storeAggregate(AggregationKey key, int amount, String vat, String posType);
+    boolean storeAggregate(AggregationKey key, long amount, String vat, String posType);
 
     /**
      * Get the aggregate computed over a single aggregation key.
